requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.3
python-dotenv==1.0.0
psycopg2-binary==2.9.11
gunicorn==21.2.0
